feat: Create theme switcher
It is important to empasize that in order to easily
change fill color of the theme switcher icon the
svg element was implemented directly into html file.
Using <img> tag didn't give option to change color
predefined in the file of the icon. This solution
may be temporary, but for now I find it the most
efficient in terms of styling.