fix: Animations and svgInject
There was a problem with injecting svg and
the solution was to put svgInject.min.js inside
of head tag.
Also tabs animations didn't adjust to window resize
so I added setTimeout to scrollTopTop script.
And last but not least animated background was
causind change f page widht so I fixed this.